+++ MiniTwit.Web/src/components/timelines/MyTimeLine.tsx
@@ -1,5 +1,5 @@
-import { getUsername } from "../authentication/JwtToken";
-import { useDocumentTitle } from "../utilities/Utilities";
+import { getUsername } from "../../authentication/JwtToken";
+import { useDocumentTitle } from "../../utilities/Utilities";
 
 function MyTimeline() {
     useDocumentTitle("My Timeline")
